Reorganize subpackages.
