+++ deps/go_test.go
@@ -39,10 +39,11 @@ func TestGo(t *testing.T) {
 		"../main.go",
 		"../main_test.go",
 		"../namer.go",
+		"../proto/deps.go",
+		"../proto/deps_test.go",
+		"../proto/rule.go",
 		"../register.go",
 		"../register_test.go",
-		"../rules/proto.go",
-		"../rules/tsdecls.go",
 		"../runner.go",
 		"../runner_test.go",
 		"../seq.go",
@@ -52,14 +53,13 @@ func TestGo(t *testing.T) {
 		"../sqlite/migrations/20220829093104_initial.sql",
 		"../target.go",
 		"../target_test.go",
+		"../ts/tsdecls.go",
 		"../types.go",
 		"../types_test.go",
 		"../yaml.go",
 		"../yaml_test.go",
 		"go.go",
 		"go_test.go",
-		"proto.go",
-		"proto_test.go",
 	}
 
 	got, err := Go(".", false)
