+++ embeds.go
@@ -2,7 +2,7 @@ package fab
 
 import "embed"
 
-//go:embed *.go go.* driver.go.tmpl deps/*.go rules/*.go sqlite/*.go sqlite/migrations/*.sql
+//go:embed *.go go.* driver.go.tmpl deps/*.go proto/*.go sqlite/*.go sqlite/migrations/*.sql ts/*.go
 var embeds embed.FS
 
 //go:embed driver.go.tmpl
