+++ proto/deps.go
@@ -1,4 +1,4 @@
-package deps
+package proto
 
 import (
 	"bufio"
@@ -16,13 +16,13 @@ import (
 	"github.com/bobg/fab"
 )
 
-// Proto reads a protocol-buffer file and returns its list of dependencies.
+// Deps reads a protocol-buffer file and returns its list of dependencies.
 // Included in the dependencies is the file itself,
 // plus any files it imports
 // (directly or indirectly)
 // that can be found among the given include directories.
 // The list is sorted for consistent, predictable results.
-func Proto(filename string, includes []string) ([]string, error) {
+func Deps(filename string, includes []string) ([]string, error) {
 	f, err := os.Open(filename)
 	if err != nil {
 		return nil, errors.Wrapf(err, "opening %s", filename)
@@ -81,11 +81,11 @@ func protodepsDecoder(node *yaml.Node) ([]string, error) {
 		Includes []string `yaml:"Includes"`
 	}
 	if err := node.Decode(&pd); err != nil {
-		return nil, errors.Wrap(err, "YAML error in deps.Proto node")
+		return nil, errors.Wrap(err, "YAML error in proto.Deps node")
 	}
-	return Proto(pd.File, pd.Includes)
+	return Deps(pd.File, pd.Includes)
 }
 
 func init() {
-	fab.RegisterYAMLStringList("deps.Proto", protodepsDecoder)
+	fab.RegisterYAMLStringList("proto.Deps", protodepsDecoder)
 }
