+++ proto/rule.go
@@ -1,4 +1,4 @@
-package rules
+package proto
 
 import (
 	"sort"
@@ -8,7 +8,6 @@ import (
 	"github.com/bobg/go-generics/v2/slices"
 
 	"github.com/bobg/fab"
-	"github.com/bobg/fab/deps"
 )
 
 // Proto produces a target that compiles protocol-buffer files using the "protoc" command.
@@ -17,11 +16,11 @@ import (
 // includes is a list of directories in which to find .proto files;
 // otherOpts are options (other than -I / --proto_path options) for the protoc command line.
 // Typically otherOpts includes at least "--foo_out=DIR" for some target language foo.
-// This function uses [deps.Proto] to find the dependencies of the input files.
+// This function uses [Deps] to find the dependencies of the input files.
 func Proto(inputs, outputs, includes, otherOpts []string) (fab.Target, error) {
 	alldeps := set.New[string](inputs...)
 	for _, inp := range inputs {
-		d, err := deps.Proto(inp, includes)
+		d, err := Deps(inp, includes)
 		if err != nil {
 			return nil, errors.Wrapf(err, "computing dependencies for %s", inp)
 		}
