+++ proto/testdata/foo.proto
